Moving DisclosureContext to Disclosure file
