+++ src/components/disclosure/components/disclosure-content/index.tsx
@@ -1,5 +1,5 @@
 import classNames from 'classnames'
-import { useDisclosure } from 'components/disclosure/provider'
+import { useDisclosureState } from 'components/disclosure'
 import disclosureStyles from 'components/disclosure/disclosure.module.css'
 import { DisclosureContentProps } from './types'
 
@@ -9,7 +9,7 @@ import { DisclosureContentProps } from './types'
  * @see https://developer.hashi-mktg.com/swingset/components/disclosure
  */
 const DisclosureContent = ({ children, className }: DisclosureContentProps) => {
-  const { contentContainerId } = useDisclosure()
+  const { contentContainerId } = useDisclosureState()
 
   return (
     <div
